Moved discord bot testing material to discord_bot_tester.py, converted discord_bot to the beginning of a wrapper for the main runner
