+++ .gitignore
@@ -7,6 +7,7 @@ CurlRequest.sh
 discord_bot_2.py
 tester.py
 *.wav
+*.mp3
 *.txt
 
 # Byte-compiled / optimized / DLL files
